// packages/sdk/src/zod/generated/AdminOrderPromotion.ts
// This file is auto-generated. Do not edit directly.
import { z } from 'zod';

export const AdminOrderPromotionSchema = z.object({
  id: z.string(),
  promotion_id: z.string(),
  name: z.string(),
  description: z.string().nullable(),
  code: z.string().nullable(),
  amount: z.string(),
  display_amount: z.string(),
});

export type AdminOrderPromotion = z.infer<typeof AdminOrderPromotionSchema>;